requirements.txt
jupyter
black
flake8
pep8-naming
flake8-annotations
mypy
isort
pytest
ipykernel
kafka-python
websocket-client
python-dotenv